+++ altair/vega/v5/__init__.py
@@ -1,4 +1,6 @@
 # flake8: noqa
+import warnings
+from ...utils.deprecation import AltairDeprecationWarning
 from .display import vega, Vega, renderers
 from .schema import *
 
@@ -11,4 +13,9 @@ from .data import (
     to_csv,
     to_values,
     default_data_transformer,
+)
+
+warnings.warn(
+    "The module altair.vega.v5 is deprecated and will be removed in Altair 5.",
+    AltairDeprecationWarning,
 )